requirements.txt
astroid==2.3.0
Flask==0.10.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==6.1.0
pkg-resources==0.0.0
py==1.8.0
pylint==2.4.1
requests==2.22.0
six==1.12.0
typed-ast==1.2.0
Werkzeug==0.16.0
wrapt==1.11.2
